Moving useResolvedNode(), adding some documentation
