+++ src/components/summary/SummaryComponentSwitch.tsx
@@ -3,12 +3,12 @@ import * as React from 'react';
 import SingleInputSummary from 'src/components/summary/SingleInputSummary';
 import SummaryBoilerplate from 'src/components/summary/SummaryBoilerplate';
 import SummaryGroupComponent from 'src/components/summary/SummaryGroupComponent';
-import { useResolvedNode } from 'src/features/expressions/useResolvedNode';
 import MultipleChoiceSummary from 'src/layout/Checkboxes/MultipleChoiceSummary';
 import { AttachmentSummaryComponent } from 'src/layout/FileUpload/AttachmentSummaryComponent';
 import { AttachmentWithTagSummaryComponent } from 'src/layout/FileUploadWithTag/AttachmentWithTagSummaryComponent';
 import HeaderSummary from 'src/layout/Header/HeaderSummary';
 import MapComponentSummary from 'src/layout/Map/MapComponentSummary';
+import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import type { ILayoutGroup } from 'src/layout/Group/types';
 import type { ILayoutComponent } from 'src/layout/layout';
 import type { ILayoutCompSummary } from 'src/layout/Summary/types';
