+++ src/features/form/containers/AltinnMobileTableItem.tsx
@@ -16,9 +16,9 @@ import { Delete as DeleteIcon, Edit as EditIcon, Warning as WarningIcon } from '
 import cn from 'classnames';
 
 import { DeleteWarningPopover } from 'src/components/molecules/DeleteWarningPopover';
-import { useResolvedNode } from 'src/features/expressions/useResolvedNode';
 import { getLanguageFromKey } from 'src/language/sharedLanguage';
 import theme from 'src/theme/altinnStudioTheme';
+import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import type { ExprResolved } from 'src/features/expressions/types';
 import type { ILayoutGroup } from 'src/layout/Group/types';
 import type { ITextResourceBindings } from 'src/types';
