+++ src/features/form/containers/DisplayGroupContainer.tsx
@@ -4,9 +4,9 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 import cn from 'classnames';
 
 import { useAppSelector } from 'src/common/hooks';
-import { useResolvedNode } from 'src/features/expressions/useResolvedNode';
 import { makeGetHidden } from 'src/selectors/getLayoutData';
 import printStyles from 'src/styles/print.module.css';
+import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import { getTextFromAppOrDefault } from 'src/utils/textResource';
 import type { ILayoutGroup } from 'src/layout/Group/types';
 import type { ILayout, ILayoutComponent, ILayoutComponentOrGroup } from 'src/layout/layout';
