+++ src/layout/GenericComponent.tsx
@@ -5,7 +5,6 @@ import { Grid, makeStyles } from '@material-ui/core';
 import classNames from 'classnames';
 
 import { useAppDispatch, useAppSelector } from 'src/common/hooks';
-import { useResolvedNode } from 'src/features/expressions/useResolvedNode';
 import Description from 'src/features/form/components/Description';
 import Label from 'src/features/form/components/Label';
 import Legend from 'src/features/form/components/Legend';
@@ -24,6 +23,7 @@ import {
   isComponentValid,
   selectComponentTexts,
 } from 'src/utils/formComponentUtils';
+import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import { renderValidationMessagesForComponent } from 'src/utils/render';
 import { getTextResourceByKey } from 'src/utils/sharedUtils';
 import type { ExprResolved } from 'src/features/expressions/types';
