+++ src/layout/Likert/RepeatingGroupsLikertContainer.tsx
@@ -6,10 +6,10 @@ import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { useAppSelector } from 'src/common/hooks';
 import { useGetOptions } from 'src/components/hooks';
 import { AltinnSpinner, AltinnTable, AltinnTableBody, AltinnTableHeader, AltinnTableRow } from 'src/components/shared';
-import { useResolvedNode } from 'src/features/expressions/useResolvedNode';
 import { GenericComponent } from 'src/layout/GenericComponent';
 import { LayoutStyle } from 'src/types';
 import { getTextResource } from 'src/utils/formComponentUtils';
+import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import { getOptionLookupKey } from 'src/utils/options';
 import type { ILayoutGroup } from 'src/layout/Group/types';
 import type { ComponentInGroup } from 'src/layout/layout';
